+++ test/sql/functions.sql
@@ -48,6 +48,12 @@ SELECT l1_distance('[0,0]', '[0,1]');
 SELECT l1_distance('[1,2]', '[3]');
 SELECT l1_distance('[3e38]', '[-3e38]');
 
+SELECT hamming_distance(B'111', B'111');
+SELECT hamming_distance(B'111', B'110');
+SELECT hamming_distance(B'111', B'100');
+SELECT hamming_distance(B'111', B'000');
+SELECT hamming_distance(B'111', B'00');
+
 SELECT avg(v) FROM unnest(ARRAY['[1,2,3]'::vector, '[3,5,7]']) v;
 SELECT avg(v) FROM unnest(ARRAY['[1,2,3]'::vector, '[3,5,7]', NULL]) v;
 SELECT avg(v) FROM unnest(ARRAY[]::vector[]) v;