+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 If upgrading with Postgres 12 or Docker, see [these notes](https://github.com/pgvector/pgvector#060).
 
 - Added support for parallel index builds for HNSW
+- Added `hamming_distance` function
 - Added validation for GUC parameters
 - Changed storage for vector from `extended` to `external`
 - Improved performance of HNSW